requirements.txt
python-crontab == 2.5; platform_system != "Windows"
